Initial mesh refactor
