+++ src/viser/client/src/OutlinesIfHovered.tsx
@@ -1,0 +1,43 @@
+import React from "react";
+import { useFrame } from "@react-three/fiber";
+import { HoverableContext } from "./HoverContext";
+import { Outlines } from "./Outlines";
+import * as THREE from "three";
+
+/** Outlines object, which should be placed as a child of all meshes that might
+ * be clickable. */
+export function OutlinesIfHovered(
+  props: { alwaysMounted?: boolean; creaseAngle?: number } = {
+    // Can be set to true for objects like meshes which may be slow to mount.
+    // It seems better to set to False for instanced meshes, there may be some
+    // drei or fiber-related race conditions...
+    alwaysMounted: false,
+    // Some thing just look better with no creasing, like camera frustum objects.
+    creaseAngle: Math.PI,
+  },
+) {
+  const groupRef = React.useRef<THREE.Group>(null);
+  const hoveredRef = React.useContext(HoverableContext);
+  const [mounted, setMounted] = React.useState(true);
+
+  useFrame(() => {
+    if (hoveredRef === null) return;
+    if (props.alwaysMounted) {
+      if (groupRef.current === null) return;
+      groupRef.current.visible = hoveredRef.current;
+    } else if (hoveredRef.current != mounted) {
+      setMounted(hoveredRef.current);
+    }
+  });
+  return hoveredRef === null || !mounted ? null : (
+    <Outlines
+      ref={groupRef}
+      thickness={10}
+      screenspace={true}
+      color={0xfbff00}
+      opacity={0.8}
+      transparent={true}
+      angle={props.creaseAngle}
+    />
+  );
+}
