+++ src/viser/client/src/ThreeAssets.tsx
@@ -1,66 +1,17 @@
-import { InstancedMesh2 } from "@three.ez/instanced-mesh";
 import { Instance, Instances, Line, shaderMaterial } from "@react-three/drei";
-import { createPortal, useFrame, useThree } from "@react-three/fiber";
-import { Outlines } from "./Outlines";
+import { useFrame, useThree } from "@react-three/fiber";
+import { OutlinesIfHovered } from "./OutlinesIfHovered";
 import React from "react";
 import { HoverableContext } from "./HoverContext";
 import * as THREE from "three";
-import { GLTF, GLTFLoader } from "three/examples/jsm/loaders/GLTFLoader";
-import {
-  MeshBasicMaterial,
-  MeshDepthMaterial,
-  MeshDistanceMaterial,
-  MeshLambertMaterial,
-  MeshMatcapMaterial,
-  MeshNormalMaterial,
-  MeshPhongMaterial,
-  MeshPhysicalMaterial,
-  MeshStandardMaterial,
-  MeshToonMaterial,
-  ShadowMaterial,
-  SpriteMaterial,
-  RawShaderMaterial,
-  ShaderMaterial,
-  PointsMaterial,
-  LineBasicMaterial,
-  LineDashedMaterial,
-} from "three";
-import { DRACOLoader } from "three/examples/jsm/loaders/DRACOLoader";
 import {
   CameraFrustumMessage,
   ImageMessage,
-  MeshMessage,
   PointCloudMessage,
-  SkinnedMeshMessage,
-  BatchedMeshesMessage,
-  GlbMessage,
-  BatchedGlbMessage,
 } from "./WebsocketMessages";
-import { ViewerContext } from "./ViewerContext";
-import { BatchedMeshManager, setupBatchedMesh } from "./BatchedMeshManager";
+import { rgbToInt } from "./mesh/MeshUtils";
+import { ShaderMaterial, MeshBasicMaterial } from "three";
 
-type AllPossibleThreeJSMaterials =
-  | MeshBasicMaterial
-  | MeshDepthMaterial
-  | MeshDistanceMaterial
-  | MeshLambertMaterial
-  | MeshMatcapMaterial
-  | MeshNormalMaterial
-  | MeshPhongMaterial
-  | MeshPhysicalMaterial
-  | MeshStandardMaterial
-  | MeshToonMaterial
-  | ShadowMaterial
-  | SpriteMaterial
-  | RawShaderMaterial
-  | ShaderMaterial
-  | PointsMaterial
-  | LineBasicMaterial
-  | LineDashedMaterial;
-
-function rgbToInt(rgb: [number, number, number]): number {
-  return (rgb[0] << 16) | (rgb[1] << 8) | rgb[2];
-}
 const originGeom = new THREE.SphereGeometry(1.0);
 
 const PointCloudMaterial = /* @__PURE__ */ shaderMaterial(
@@ -196,277 +147,6 @@ export const PointCloud = React.forwardRef<THREE.Points, PointCloudMessage>(
     return <points ref={ref} geometry={geometry} material={material} />;
   },
 );
-
-/** Component for rendering the contents of GLB files. */
-export const GlbAsset = React.forwardRef<
-  THREE.Group,
-  GlbMessage | BatchedGlbMessage
->(function GlbAsset(message, ref) {
-  const props = message.props;
-
-  // Create persistent geometry and material. Set attributes when we receive updates.
-  const [gltf, setGltf] = React.useState<GLTF>();
-  const [meshes, setMeshes] = React.useState<THREE.Mesh[]>([]);
-
-  // glTF/GLB files support animations.
-  const mixerRef = React.useRef<THREE.AnimationMixer | null>(null);
-  React.useEffect(() => {
-    const loader = new GLTFLoader();
-
-    // We use a CDN for Draco. We could move this locally if we want to use Viser offline.
-    const dracoLoader = new DRACOLoader();
-    dracoLoader.setDecoderPath("https://www.gstatic.com/draco/v1/decoders/");
-    loader.setDRACOLoader(dracoLoader);
-
-    const glb_data = new Uint8Array(props.glb_data);
-    loader.parse(
-      glb_data.buffer,
-      "",
-      (gltf) => {
-        if (gltf.animations && gltf.animations.length) {
-          mixerRef.current = new THREE.AnimationMixer(gltf.scene);
-          gltf.animations.forEach((clip) => {
-            mixerRef.current!.clipAction(clip).play();
-          });
-        }
-        const meshes: THREE.Mesh[] = [];
-        gltf?.scene.traverse((obj) => {
-          if (obj instanceof THREE.Mesh) {
-            obj.geometry.computeVertexNormals();
-            obj.geometry.computeBoundingSphere();
-            obj.castShadow = props.cast_shadow;
-            obj.receiveShadow = props.receive_shadow;
-            meshes.push(obj);
-          }
-        });
-
-        setMeshes(meshes);
-        setGltf(gltf);
-      },
-      (error) => {
-        console.log("Error loading GLB!");
-        console.log(error);
-      },
-    );
-
-    return () => {
-      if (mixerRef.current) mixerRef.current.stopAllAction();
-
-      function disposeNode(node: any) {
-        if (node instanceof THREE.Mesh) {
-          if (node.geometry) {
-            node.geometry.dispose();
-          }
-          if (node.material) {
-            if (Array.isArray(node.material)) {
-              node.material.forEach((material) => {
-                disposeMaterial(material);
-              });
-            } else {
-              disposeMaterial(node.material);
-            }
-          }
-        }
-      }
-      function disposeMaterial(material: AllPossibleThreeJSMaterials) {
-        if ("map" in material) material.map?.dispose();
-        if ("lightMap" in material) material.lightMap?.dispose();
-        if ("bumpMap" in material) material.bumpMap?.dispose();
-        if ("normalMap" in material) material.normalMap?.dispose();
-        if ("specularMap" in material) material.specularMap?.dispose();
-        if ("envMap" in material) material.envMap?.dispose();
-        if ("alphaMap" in material) material.alphaMap?.dispose();
-        if ("aoMap" in material) material.aoMap?.dispose();
-        if ("displacementMap" in material) material.displacementMap?.dispose();
-        if ("emissiveMap" in material) material.emissiveMap?.dispose();
-        if ("gradientMap" in material) material.gradientMap?.dispose();
-        if ("metalnessMap" in material) material.metalnessMap?.dispose();
-        if ("roughnessMap" in material) material.roughnessMap?.dispose();
-        material.dispose(); // disposes any programs associated with the material
-      }
-
-      // Attempt to free resources.
-      gltf?.scene.traverse(disposeNode);
-    };
-  }, [props.glb_data, props.cast_shadow, props.receive_shadow]);
-
-  useFrame((_, delta) => {
-    if (mixerRef.current) {
-      mixerRef.current.update(delta);
-    }
-  });
-
-  // Create the instanced meshes for batched GLBs:
-  // - Flattening the hierarchy to store root-to-node transformations, then
-  // - Applying both the root-to-world and node-to-root transformations.
-  // We additionally generate LODs for each mesh, if requested.
-  const [transforms, setTransforms] = React.useState<
-    {
-      position: THREE.Vector3;
-      rotation: THREE.Quaternion;
-      scale: THREE.Vector3;
-    }[]
-  >([]);
-
-  // Use a ref to store mesh managers for proper disposal
-  const meshManagersRef = React.useRef<{
-    gltfScene: THREE.Group;
-    managers: BatchedMeshManager[];
-  } | null>(null);
-
-  // Initialize mesh managers
-  React.useEffect(() => {
-    if (message.type !== "BatchedGlbMessage" || !gltf) return;
-
-    // Clean up previous managers if they exist
-    if (meshManagersRef.current) {
-      meshManagersRef.current.managers.forEach((manager) => {
-        manager.dispose();
-      });
-    }
-
-    const scene = gltf.scene.clone();
-    const managers: BatchedMeshManager[] = [];
-    const transforms: {
-      position: THREE.Vector3;
-      rotation: THREE.Quaternion;
-      scale: THREE.Vector3;
-    }[] = [];
-
-    scene.traverse((node) => {
-      if (node instanceof THREE.Mesh && node.parent) {
-        // Store transform info.
-        const position = new THREE.Vector3();
-        const scale = new THREE.Vector3();
-        const quat = new THREE.Quaternion();
-        node.getWorldPosition(position);
-        node.getWorldScale(scale);
-        node.getWorldQuaternion(quat);
-
-        const transform = {
-          position: position.clone(),
-          rotation: quat.clone(),
-          scale: scale.clone(),
-        };
-        transforms.push(transform);
-
-        // Create instanced mesh with LOD
-        const numInstances =
-          message.props.batched_positions.byteLength /
-          (3 * Float32Array.BYTES_PER_ELEMENT);
-        const manager = setupBatchedMesh(
-          node.geometry.clone(),
-          node.material,
-          numInstances,
-          message.props.lod,
-          message.props.cast_shadow,
-          Math.max(scale.x, scale.y, scale.z),
-        );
-
-        // Hide the original node.
-        node.visible = false;
-
-        // Add the instanced mesh to the scene.
-        managers.push(manager);
-        scene.add(manager.getMesh());
-      }
-    });
-
-    setTransforms(transforms);
-    meshManagersRef.current = { gltfScene: scene, managers };
-
-    // Clean up when component unmounts or dependencies change
-    return () => {
-      if (meshManagersRef.current) {
-        meshManagersRef.current.managers.forEach((manager) => {
-          manager.dispose();
-        });
-        meshManagersRef.current = null;
-      }
-    };
-  }, [
-    message.type,
-    gltf,
-    ...(message.type === "BatchedGlbMessage"
-      ? [message.props.lod, message.props.cast_shadow]
-      : []),
-  ]);
-
-  // Handle updates to instance positions/orientations.
-  React.useEffect(() => {
-    if (message.type !== "BatchedGlbMessage" || !meshManagersRef.current)
-      return;
-
-    const batched_positions = new Float32Array(
-      message.props.batched_positions.buffer.slice(
-        message.props.batched_positions.byteOffset,
-        message.props.batched_positions.byteOffset +
-          message.props.batched_positions.byteLength,
-      ),
-    );
-
-    const batched_wxyzs = new Float32Array(
-      message.props.batched_wxyzs.buffer.slice(
-        message.props.batched_wxyzs.byteOffset,
-        message.props.batched_wxyzs.byteOffset +
-          message.props.batched_wxyzs.byteLength,
-      ),
-    );
-
-    // Update instance count if needed.
-    const newNumInstances =
-      message.props.batched_positions.byteLength /
-      (3 * Float32Array.BYTES_PER_ELEMENT);
-
-    // Update all mesh managers
-    meshManagersRef.current.managers.forEach((manager, mesh_index) => {
-      manager.setInstanceCount(newNumInstances);
-      manager.updateInstances(
-        batched_positions,
-        batched_wxyzs,
-        transforms[mesh_index],
-      );
-    });
-  }, [
-    message.type,
-    transforms,
-    ...(message.type === "BatchedGlbMessage"
-      ? [message.props.batched_positions, message.props.batched_wxyzs]
-      : []),
-  ]);
-
-  if (!gltf) return null;
-
-  return (
-    <group ref={ref}>
-      {message.type === "BatchedGlbMessage" ? (
-        meshManagersRef.current && (
-          <>
-            <primitive
-              object={meshManagersRef.current.gltfScene}
-              scale={message.props.scale}
-            />
-            <OutlinesIfHovered alwaysMounted={false} />
-          </>
-        )
-      ) : (
-        <>
-          <primitive
-            object={gltf.scene}
-            scale={message.props.scale}
-            castShadow={message.props.cast_shadow}
-            receiveShadow={message.props.receive_shadow}
-          />
-          {meshes.map((mesh, i) => (
-            <React.Fragment key={i}>
-              {createPortal(<OutlinesIfHovered alwaysMounted />, mesh)}
-            </React.Fragment>
-          ))}
-        </>
-      )}
-    </group>
-  );
-});
 
 /** Helper for adding coordinate frames as scene nodes. */
 export const CoordinateFrame = React.forwardRef<
@@ -623,379 +303,6 @@ export const InstancedAxes = React.forwardRef<
   );
 });
 
-/** Convert raw RGB color buffers to linear color buffers. **/
-export const ViserMesh = React.forwardRef<
-  THREE.Mesh | THREE.SkinnedMesh | InstancedMesh2,
-  MeshMessage | SkinnedMeshMessage | BatchedMeshesMessage
->(function ViserMesh(message, ref) {
-  const viewer = React.useContext(ViewerContext)!;
-
-  const generateGradientMap = (shades: 3 | 5) => {
-    const texture = new THREE.DataTexture(
-      Uint8Array.from(shades == 3 ? [0, 128, 255] : [0, 64, 128, 192, 255]),
-      shades,
-      1,
-      THREE.RedFormat,
-    );
-
-    texture.needsUpdate = true;
-    return texture;
-  };
-
-  const assertUnreachable = (x: never): never => {
-    throw new Error(`Should never get here! ${x}`);
-  };
-
-  const [material, setMaterial] = React.useState<THREE.Material>();
-  const bonesRef = React.useRef<THREE.Bone[]>();
-
-  React.useEffect(() => {
-    const standardArgs = {
-      color:
-        message.props.color === null
-          ? undefined
-          : rgbToInt(message.props.color),
-      wireframe: message.props.wireframe,
-      transparent: message.props.opacity !== null,
-      opacity: message.props.opacity ?? 1.0,
-      // Flat shading only makes sense for non-wireframe materials.
-      flatShading: message.props.flat_shading && !message.props.wireframe,
-      side: {
-        front: THREE.FrontSide,
-        back: THREE.BackSide,
-        double: THREE.DoubleSide,
-      }[message.props.side],
-    };
-    const material =
-      message.props.material == "standard" || message.props.wireframe
-        ? new THREE.MeshStandardMaterial(standardArgs)
-        : message.props.material == "toon3"
-          ? new THREE.MeshToonMaterial({
-              gradientMap: generateGradientMap(3),
-              ...standardArgs,
-            })
-          : message.props.material == "toon5"
-            ? new THREE.MeshToonMaterial({
-                gradientMap: generateGradientMap(5),
-                ...standardArgs,
-              })
-            : assertUnreachable(message.props.material);
-    setMaterial(material);
-
-    return () => {
-      // Dispose material when done.
-      material.dispose();
-    };
-  }, [
-    message.props.material,
-    message.props.color,
-    message.props.wireframe,
-    message.props.opacity,
-  ]);
-
-  // Create persistent geometry. Set attributes when we receive updates.
-  const [geometry, setGeometry] = React.useState<THREE.BufferGeometry>();
-  const [skeleton, setSkeleton] = React.useState<THREE.Skeleton>();
-
-  // Setup geometry attributes
-  React.useEffect(() => {
-    const geometry = new THREE.BufferGeometry();
-    geometry.setAttribute(
-      "position",
-      new THREE.BufferAttribute(
-        new Float32Array(
-          message.props.vertices.buffer.slice(
-            message.props.vertices.byteOffset,
-            message.props.vertices.byteOffset +
-              message.props.vertices.byteLength,
-          ),
-        ),
-        3,
-      ),
-    );
-    geometry.setIndex(
-      new THREE.BufferAttribute(
-        new Uint32Array(
-          message.props.faces.buffer.slice(
-            message.props.faces.byteOffset,
-            message.props.faces.byteOffset + message.props.faces.byteLength,
-          ),
-        ),
-        1,
-      ),
-    );
-    geometry.computeVertexNormals();
-    geometry.computeBoundingSphere();
-
-    // Handle skinned mesh setup if needed
-    let skeleton = undefined;
-    if (message.type === "SkinnedMeshMessage") {
-      // Skinned mesh.
-      const bone_wxyzs = new Float32Array(
-        message.props.bone_wxyzs.buffer.slice(
-          message.props.bone_wxyzs.byteOffset,
-          message.props.bone_wxyzs.byteOffset +
-            message.props.bone_wxyzs.byteLength,
-        ),
-      );
-      const bone_positions = new Float32Array(
-        message.props.bone_positions.buffer.slice(
-          message.props.bone_positions.byteOffset,
-          message.props.bone_positions.byteOffset +
-            message.props.bone_positions.byteLength,
-        ),
-      );
-
-      const bones: THREE.Bone[] = [];
-      bonesRef.current = bones;
-      for (let i = 0; i < bone_positions.length / 3; i++) {
-        bones.push(new THREE.Bone());
-      }
-      const boneInverses: THREE.Matrix4[] = [];
-      const xyzw_quat = new THREE.Quaternion();
-      bones.forEach((bone, i) => {
-        xyzw_quat.set(
-          bone_wxyzs[i * 4 + 1],
-          bone_wxyzs[i * 4 + 2],
-          bone_wxyzs[i * 4 + 3],
-          bone_wxyzs[i * 4 + 0],
-        );
-
-        const boneInverse = new THREE.Matrix4();
-        boneInverse.makeRotationFromQuaternion(xyzw_quat);
-        boneInverse.setPosition(
-          bone_positions[i * 3 + 0],
-          bone_positions[i * 3 + 1],
-          bone_positions[i * 3 + 2],
-        );
-        boneInverse.invert();
-        boneInverses.push(boneInverse);
-
-        bone.setRotationFromQuaternion(xyzw_quat);
-        bone.position.set(
-          bone_positions[i * 3 + 0],
-          bone_positions[i * 3 + 1],
-          bone_positions[i * 3 + 2],
-        );
-      });
-      skeleton = new THREE.Skeleton(bones, boneInverses);
-
-      geometry.setAttribute(
-        "skinIndex",
-        new THREE.BufferAttribute(
-          new Uint16Array(
-            message.props.skin_indices.buffer.slice(
-              message.props.skin_indices.byteOffset,
-              message.props.skin_indices.byteOffset +
-                message.props.skin_indices.byteLength,
-            ),
-          ),
-          4,
-        ),
-      );
-      geometry.setAttribute(
-        "skinWeight",
-        new THREE.BufferAttribute(
-          new Float32Array(
-            message.props.skin_weights!.buffer.slice(
-              message.props.skin_weights!.byteOffset,
-              message.props.skin_weights!.byteOffset +
-                message.props.skin_weights!.byteLength,
-            ),
-          ),
-          4,
-        ),
-      );
-    }
-    skeleton?.init();
-    setGeometry(geometry);
-    setSkeleton(skeleton);
-    return () => {
-      if (message.type === "SkinnedMeshMessage") {
-        skeleton !== undefined && skeleton.dispose();
-        const state = viewer.skinnedMeshState.current[message.name];
-        state.initialized = false;
-      }
-    };
-  }, [
-    message.type,
-    message.props.vertices.buffer,
-    message.props.faces.buffer,
-    message.type == "SkinnedMeshMessage"
-      ? message.props.skin_indices.buffer
-      : null,
-    message.type == "SkinnedMeshMessage"
-      ? message.props.skin_weights.buffer
-      : null,
-    message.type == "SkinnedMeshMessage"
-      ? message.props.bone_wxyzs.buffer
-      : null,
-    message.type == "SkinnedMeshMessage"
-      ? message.props.bone_positions.buffer
-      : null,
-  ]);
-  // Dispose geometry when done.
-  React.useEffect(() => {
-    return () => {
-      geometry !== undefined && geometry.dispose();
-    };
-  }, [geometry]);
-
-  // Update bone transforms for skinned meshes.
-  useFrame(() => {
-    if (message.type !== "SkinnedMeshMessage") return;
-
-    const parentNode = viewer.nodeRefFromName.current[message.name];
-    if (parentNode === undefined) return;
-
-    const state = viewer.skinnedMeshState.current[message.name];
-    const bones = bonesRef.current;
-    if (skeleton !== undefined && bones !== undefined) {
-      if (!state.initialized) {
-        bones.forEach((bone) => {
-          parentNode.add(bone);
-        });
-        state.initialized = true;
-      }
-      bones.forEach((bone, i) => {
-        const wxyz = state.poses[i].wxyz;
-        const position = state.poses[i].position;
-        bone.quaternion.set(wxyz[1], wxyz[2], wxyz[3], wxyz[0]);
-        bone.position.set(position[0], position[1], position[2]);
-      });
-    }
-  });
-
-  // Create the batched mesh manager once
-  const meshManagerRef = React.useRef<BatchedMeshManager | null>(null);
-
-  // Create the instanced mesh once
-  React.useEffect(() => {
-    if (
-      message.type !== "BatchedMeshesMessage" ||
-      material === undefined ||
-      geometry === undefined
-    )
-      return;
-
-    const numInstances =
-      message.props.batched_positions.byteLength /
-      (3 * Float32Array.BYTES_PER_ELEMENT);
-
-    // Create new manager
-    meshManagerRef.current = setupBatchedMesh(
-      geometry,
-      material,
-      numInstances,
-      message.props.lod,
-      message.props.cast_shadow,
-    );
-
-    // Cleanup function to ensure proper disposal
-    return () => {
-      if (meshManagerRef.current) {
-        meshManagerRef.current.dispose();
-        meshManagerRef.current = null;
-      }
-    };
-  }, [
-    message.type,
-    material,
-    geometry,
-    ...(message.type === "BatchedMeshesMessage"
-      ? [message.props.lod, message.props.cast_shadow]
-      : []),
-  ]);
-
-  // Handle updates to instance positions/orientations
-  React.useEffect(() => {
-    if (message.type !== "BatchedMeshesMessage" || !meshManagerRef.current)
-      return;
-
-    const batched_positions = new Float32Array(
-      message.props.batched_positions.buffer.slice(
-        message.props.batched_positions.byteOffset,
-        message.props.batched_positions.byteOffset +
-          message.props.batched_positions.byteLength,
-      ),
-    );
-
-    const batched_wxyzs = new Float32Array(
-      message.props.batched_wxyzs.buffer.slice(
-        message.props.batched_wxyzs.byteOffset,
-        message.props.batched_wxyzs.byteOffset +
-          message.props.batched_wxyzs.byteLength,
-      ),
-    );
-
-    // Update instance count if needed
-    const newNumInstances =
-      message.props.batched_positions.byteLength /
-      (3 * Float32Array.BYTES_PER_ELEMENT);
-    meshManagerRef.current.setInstanceCount(newNumInstances);
-
-    // Update instance transforms
-    meshManagerRef.current.updateInstances(batched_positions, batched_wxyzs);
-  }, [
-    message.type,
-    ...(message.type === "BatchedMeshesMessage"
-      ? [
-          message.props.batched_positions.buffer,
-          message.props.batched_wxyzs.buffer,
-        ]
-      : []),
-  ]);
-
-  if (geometry === undefined || material === undefined) {
-    return null;
-  }
-
-  // Render the appropriate mesh type
-  if (message.type === "BatchedMeshesMessage") {
-    return (
-      <>
-        {meshManagerRef.current && (
-          <>
-            <primitive object={meshManagerRef.current.getMesh()} />
-            <OutlinesIfHovered alwaysMounted />
-          </>
-        )}
-      </>
-    );
-  } else if (message.type === "SkinnedMeshMessage") {
-    return (
-      <skinnedMesh
-        ref={ref as React.ForwardedRef<THREE.SkinnedMesh>}
-        geometry={geometry}
-        material={material}
-        skeleton={skeleton}
-        castShadow={message.props.cast_shadow}
-        receiveShadow={message.props.receive_shadow}
-        // TODO: leaving culling on (default) sometimes causes the
-        // mesh to randomly disappear, as of r3f==8.16.2.
-        //
-        // Probably this is because we don't update the bounding
-        // sphere after the bone transforms change.
-        frustumCulled={false}
-      >
-        <OutlinesIfHovered alwaysMounted />
-      </skinnedMesh>
-    );
-  } else {
-    return (
-      <mesh
-        ref={ref as React.ForwardedRef<THREE.Mesh>}
-        geometry={geometry}
-        material={material}
-        castShadow={message.props.cast_shadow}
-        receiveShadow={message.props.receive_shadow}
-      >
-        <OutlinesIfHovered alwaysMounted />
-      </mesh>
-    );
-  }
-});
-
 export const ViserImage = React.forwardRef<THREE.Group, ImageMessage>(
   function ViserImage(message, ref) {
     const [imageTexture, setImageTexture] = React.useState<THREE.Texture>();
@@ -1139,41 +446,3 @@ export const CameraFrustum = React.forwardRef<
     </group>
   );
 });
-
-/** Outlines object, which should be placed as a child of all meshes that might
- * be clickable. */
-export function OutlinesIfHovered(
-  props: { alwaysMounted?: boolean; creaseAngle?: number } = {
-    // Can be set to true for objects like meshes which may be slow to mount.
-    // It seems better to set to False for instanced meshes, there may be some
-    // drei or fiber-related race conditions...
-    alwaysMounted: false,
-    // Some thing just look better with no creasing, like camera frustum objects.
-    creaseAngle: Math.PI,
-  },
-) {
-  const groupRef = React.useRef<THREE.Group>(null);
-  const hoveredRef = React.useContext(HoverableContext);
-  const [mounted, setMounted] = React.useState(true);
-
-  useFrame(() => {
-    if (hoveredRef === null) return;
-    if (props.alwaysMounted) {
-      if (groupRef.current === null) return;
-      groupRef.current.visible = hoveredRef.current;
-    } else if (hoveredRef.current != mounted) {
-      setMounted(hoveredRef.current);
-    }
-  });
-  return hoveredRef === null || !mounted ? null : (
-    <Outlines
-      ref={groupRef}
-      thickness={10}
-      screenspace={true}
-      color={0xfbff00}
-      opacity={0.8}
-      transparent={true}
-      angle={props.creaseAngle}
-    />
-  );
-}
